tx inputs as set
