+++ rust/src/builders/voting_builder.rs
@@ -146,7 +146,7 @@ impl VotingBuilder {
                 None => {}
             }
         }
-        TransactionInputs(inputs)
+        TransactionInputs::from_vec(inputs)
     }
 
     pub fn get_native_scripts(&self) -> NativeScripts {
