+++ rust/src/builders/withdrawals_builder.rs
@@ -121,7 +121,7 @@ impl WithdrawalsBuilder {
                 None => {}
             }
         }
-        TransactionInputs(inputs)
+        TransactionInputs::from_vec(inputs)
     }
 
     pub fn get_native_scripts(&self) -> NativeScripts {
